+++ app/src/main/java/com/pmd/rentavehiculos/data/model/RentarVehiculosResponse.kt
@@ -3,6 +3,6 @@ package com.pmd.rentavehiculos.data.model
 data class RentarVehiculoResponse(
     val success: Boolean,
     val message: String,
-    val rentalId: Int?,
+    val rentalId: Double?,
     val totalPrice: Double?
 )
